requirements.txt
email-validator>=2.0.0,<2.1.0
fastapi>=0.115.2,<0.116.0
httpx>=0.28.1,<0.30.0
loguru>=0.7.3,<0.8.0
pandas>=2.2.3,<2.3.0
passlib>=1.7.4,<1.8.0
pytest>=8.3.5,<9.0.0
python-dotenv>=1.1.0,<1.2.0
python-jose[cryptography]>=3.4.0,<3.5.0
python-multipart>=0.0.20,<0.0.30
uvicorn>=0.34.0,<0.40.0